+++ src/other/utils.ts
@@ -2,6 +2,7 @@ import * as path from 'path';
 import * as fs from 'fs';
 import glob = require('glob');
 import { ModFolder } from './modFolder';
+import * as child from 'child_process';
 
 export function ensureDir(path: string) {
   if (!fs.existsSync(path)) {
@@ -145,16 +146,18 @@ export function getRequiredLoadAfterIds(modinfo: any): string[] {
 }
 
 export function readModinfo(modPath: string): any {
+  let result;
   try {
     if (fs.existsSync(path.join(modPath, 'modinfo.json'))) {
-      return { 
+      result = {
         'modinfo': JSON.parse(fs.readFileSync(path.join(modPath, 'modinfo.json'), 'utf8')),
         getRequiredLoadAfterIds
       };
     }
     else if (fs.existsSync(path.join(modPath, 'annomod.json'))) {
-      return {
+      result = {
         ...JSON.parse(fs.readFileSync(path.join(modPath, 'annomod.json'), 'utf8')),
+        'modinfo': JSON.parse(fs.readFileSync(path.join(modPath, 'annomod.json'), 'utf8')),
         getRequiredLoadAfterIds
       };
     }
@@ -163,6 +166,9 @@ export function readModinfo(modPath: string): any {
     // mod jsons can be invalid pretty fast
     return undefined;
   }
+
+  result.src = result.src ?? ".";
+
   return undefined;
 }
 
@@ -303,4 +309,39 @@ export function hasGraphicsFile(modPaths: string[], filePath: string, annoRda?: 
   }
 
   return checked;
+}
+
+export interface ILogger {
+  log: (text: string) => void;
+  warn: (text: string) => void;
+  error: (text: string) => void;
+}
+
+export function downloadFile(sourceUrl: string, targetPath: string, logger?: ILogger) {
+  ensureDir(path.dirname(targetPath));
+  try {
+    child.execFileSync('curl', [
+      '-L',
+      '-o', targetPath,
+      sourceUrl
+    ]);
+  }
+  catch (e) {
+    logger?.error((<Error>e).message);
+    throw e;
+  }
+}
+
+export function extractZip(sourceZipPath: string, targetPath: string, logger?: ILogger) {
+  ensureDir(path.dirname(targetPath));
+  try {
+    child.execFileSync('tar', [
+      '-xf', sourceZipPath,
+      '-C', targetPath
+    ]);
+  }
+  catch (e) {
+    logger?.error((<Error>e).message);
+    throw e;
+  }
 }